+++ ui/src/app/app.routes.ts
@@ -20,14 +20,6 @@ export const routes: Routes = [{
                     return null
                 }, loadComponent: () => import('./pages/files/files.component').then(m => m.FilesComponent)
             },
-            {
-                matcher: (url) => {
-                    if (url.length > 0 && url[0].path === 'preview') {
-                        return { consumed: url, posParams: {} }
-                    }
-                    return null
-                }, loadComponent: () => import('./pages/file-editor/file-editor.component').then(m => m.FileEditorComponent)
-            },
             { path: 'images', loadComponent: () => import('./pages/images/images.component').then(m => m.ImagesComponent) },
             { path: 'images/:id', loadComponent: () => import('./pages/image-inspection/image-inspection.component').then(m => m.ImageInspectionComponent) },
             { path: 'logs/:id', loadComponent: () => import('./pages/logs/logs.component').then(m => m.LogsComponent) },
